(iOS) Use newer API when UIScenes are being used
This has no effect for current Cordova versions using the default
template, but future versions of Cordova iOS are likely to use UIScenes
and that may already be true for apps embedding Cordova as a library.

Partially closes #16 (although the deprecated API is still used, and now
Apple has deprecated all of SKStoreReviewController with no replacement
in favour of a SwiftUI-only implementation).